+++ src/PolyrhythmClock.cpp
@@ -20,7 +20,14 @@ struct PolyrhythmClock : Module {
         NUM_PARAMS = TUPLETS_RAND_PARAM + 3
     };
     enum InputIds {
+        RESET_INPUT,
         EXT_CLOCK_INPUT,
+        TUPLET1_RHYTHM_INPUT,
+        TUPLET1_DUR_INPUT,
+        TUPLET2_RHYTHM_INPUT,
+        TUPLET2_DUR_INPUT,
+        TUPLET3_RHYTHM_INPUT,
+        TUPLET3_DUR_INPUT,
         NUM_INPUTS
     };
     enum OutputIds {
@@ -35,7 +42,7 @@ struct PolyrhythmClock : Module {
         NUM_LIGHTS
     };
 
-    dsp::SchmittTrigger toggleTrig, bpmInputTrig;
+    dsp::SchmittTrigger toggleTrig, bpmInputTrig, resetTrig;
     dsp::PulseGenerator gatePulses[4];
     bool tupletGates[4] = {};
     bool clockOn = false;
@@ -54,17 +61,23 @@ struct PolyrhythmClock : Module {
     float phaseTuplet1 = 0;
     float phaseTuplet2 = 0;
     float phaseTuplet3 = 0;
+    float rhythm1;
+    float rhythm2;
+    float rhythm3;
+    float dur1;
+    float dur2;
+    float dur3;
 
     PolyrhythmClock() {
         config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
         configParam(CLOCK_TOGGLE_PARAM, 0.0, 1.0, 0.0, "toggle clock");
         configParam(BPM_PARAM, -2.0, 6.0, 1.0, "Tempo", " bpm", 2.0, 60.0);
-        configParam(TUPLET1_RHYTHM_PARAM, 1.0, 13.0, 1.0);
-        configParam(TUPLET1_DUR_PARAM, 1.0, 13.0, 1.0);
-        configParam(TUPLET2_RHYTHM_PARAM, 1.0, 13.0, 1.0);
-        configParam(TUPLET2_DUR_PARAM, 1.0, 13.0, 1.0);
-        configParam(TUPLET3_RHYTHM_PARAM, 1.0, 13.0, 1.0);
-        configParam(TUPLET3_DUR_PARAM, 1.0, 13.0, 1.0);
+        configParam(TUPLET1_RHYTHM_PARAM, 1.0, 24.0, 1.0);
+        configParam(TUPLET1_DUR_PARAM, 1.0, 24.0, 1.0);
+        configParam(TUPLET2_RHYTHM_PARAM, 1.0, 24.0, 1.0);
+        configParam(TUPLET2_DUR_PARAM, 1.0, 24.0, 1.0);
+        configParam(TUPLET3_RHYTHM_PARAM, 1.0, 24.0, 1.0);
+        configParam(TUPLET3_DUR_PARAM, 1.0, 24.0, 1.0);
 
         currentRhythmFraction[0] = params[TUPLET1_RHYTHM_PARAM].getValue() / params[TUPLET1_DUR_PARAM].getValue();
         currentRhythmFraction[1] = params[TUPLET2_RHYTHM_PARAM].getValue() / params[TUPLET2_DUR_PARAM].getValue();
@@ -118,8 +131,54 @@ struct PolyrhythmClock : Module {
 
     void process(const ProcessArgs& args) override {
 
+        if (resetTrig.process(inputs[RESET_INPUT].getVoltage())) {
+            resetPhases();
+        }
+
         if (toggleTrig.process(params[CLOCK_TOGGLE_PARAM].getValue())) {
             clockOn = !clockOn;
+        }
+
+        // fraction 1
+        if (inputs[TUPLET1_RHYTHM_INPUT].isConnected()) {
+            rhythm1 = inputs[TUPLET1_RHYTHM_INPUT].getVoltage() * 12.0;
+            rhythm1 = clamp(rhythm1, 0.f, 24.f);
+        } else {
+            rhythm1 = params[TUPLET1_RHYTHM_PARAM].getValue();
+        }
+        if (inputs[TUPLET1_DUR_INPUT].isConnected()) {
+            dur1 = inputs[TUPLET1_DUR_INPUT].getVoltage() * 12.0;
+            dur1 = clamp(dur1, 1.f, 24.f);
+        } else {
+            dur1 = params[TUPLET1_DUR_PARAM].getValue();
+        }
+
+        // fraction 2
+        if (inputs[TUPLET2_RHYTHM_INPUT].isConnected()) {
+            rhythm2 = inputs[TUPLET2_RHYTHM_INPUT].getVoltage() * 12.0;
+            rhythm2 = clamp(rhythm2, 0.f, 24.f);
+        } else {
+            rhythm2 = params[TUPLET2_RHYTHM_PARAM].getValue();
+        }
+        if (inputs[TUPLET2_DUR_INPUT].isConnected()) {
+            dur2 = inputs[TUPLET2_DUR_INPUT].getVoltage() * 12.0;
+            dur2 = clamp(dur2, 1.f, 24.f);
+        } else {
+            dur2 = params[TUPLET2_DUR_PARAM].getValue();
+        }
+
+        // fraction 3
+        if (inputs[TUPLET3_RHYTHM_INPUT].isConnected()) {
+            rhythm3 = inputs[TUPLET3_RHYTHM_INPUT].getVoltage() * 12.0;
+            rhythm3 = clamp(rhythm3, 0.f, 24.f);
+        } else {
+            rhythm3 = params[TUPLET3_RHYTHM_PARAM].getValue();
+        }
+        if (inputs[TUPLET3_DUR_INPUT].isConnected()) {
+            dur3 = inputs[TUPLET3_DUR_INPUT].getVoltage() * 12.0;
+            dur3 = clamp(dur3, 1.f, 24.f);
+        } else {
+            dur3 = params[TUPLET3_DUR_PARAM].getValue();
         }
 
         lights[TOGGLE_LIGHT].setBrightness(clockOn ? 1.0 : 0.0);
@@ -165,9 +224,9 @@ struct PolyrhythmClock : Module {
                 }
             }
 
-            float frac1 = params[TUPLET1_RHYTHM_PARAM].getValue() / params[TUPLET1_DUR_PARAM].getValue();
-            float frac2 = params[TUPLET2_RHYTHM_PARAM].getValue() / params[TUPLET2_DUR_PARAM].getValue();
-            float frac3 = params[TUPLET3_RHYTHM_PARAM].getValue() / params[TUPLET3_DUR_PARAM].getValue();
+            float frac1 = rhythm1 / dur1;
+            float frac2 = rhythm2 / dur2;
+            float frac3 = rhythm3 / dur3;
             if (currentRhythmFraction[0] != frac1 && phases[0] == 0.0) {
                 currentRhythmFraction[0] = frac1;
                 phases[1] = 0.0;
@@ -270,8 +329,8 @@ struct RatioDisplay : Widget {
         // nvgFill(args.vg);
         if (module == NULL) return;
 
-        int num1 = (int)module->params[PolyrhythmClock::TUPLET1_RHYTHM_PARAM].getValue();
-        int den1 = (int)module->params[PolyrhythmClock::TUPLET1_DUR_PARAM].getValue();
+        int num1 = (int)module->rhythm1;
+        int den1 = (int)module->dur1;
         text1 = std::to_string(num1) + ":" + std::to_string(den1);
         float xPos1 = num1 < 10 ? 7.6 : 0.0;
         nvgTextAlign(args.vg, NVG_ALIGN_LEFT + NVG_ALIGN_TOP);
@@ -281,8 +340,8 @@ struct RatioDisplay : Widget {
         nvgFontSize(args.vg, fontSize);
 		nvgText(args.vg, xPos1, 0, text1.c_str(), NULL);
 
-        int num2 = (int)module->params[PolyrhythmClock::TUPLET2_RHYTHM_PARAM].getValue();
-        int den2 = (int)module->params[PolyrhythmClock::TUPLET2_DUR_PARAM].getValue();
+        int num2 = (int)module->rhythm2;
+        int den2 = (int)module->dur2;
         text2 = std::to_string(num2) + ":" + std::to_string(den2);
         float xPos2 = num2 < 10 ? 7.6 : 0.0;
         // nvgTextAlign(args.vg, NVG_ALIGN_CENTER);
@@ -292,8 +351,8 @@ struct RatioDisplay : Widget {
         // nvgText(args.vg, 5, 5, text2.c_str(), NULL);
         nvgText(args.vg, xPos2, 75.4, text2.c_str(), NULL);
 
-        int num3 = (int)module->params[PolyrhythmClock::TUPLET3_RHYTHM_PARAM].getValue();
-        int den3 = (int)module->params[PolyrhythmClock::TUPLET3_DUR_PARAM].getValue();
+        int num3 = (int)module->rhythm3;
+        int den3 = (int)module->dur3;
         text3 = std::to_string(num3) + ":" + std::to_string(den3);
         float xPos3 = num3 < 10 ? 7.6 : 0.0;
         // nvgTextAlign(args.vg, NVG_ALIGN_CENTER);
@@ -338,12 +397,20 @@ struct PolyrhythmClockWidget : ModuleWidget {
         addChild(createWidget<JeremyScrew>(Vec(box.size.x-12-12, 2)));
         addChild(createWidget<JeremyScrew>(Vec(box.size.x-12-12, box.size.y - 14)));
 
-        addChild(createLight<SmallLight<JeremyAquaLight>>(Vec(29.8 - 3.21, 54 - 3.21), module, PolyrhythmClock::TOGGLE_LIGHT));
+        // addChild(createLight<SmallLight<JeremyAquaLight>>(Vec(29.8 - 3.21, 54 - 3.21), module, PolyrhythmClock::TOGGLE_LIGHT));
 
-        addParam(createParamCentered<TinyPurpleButton>(Vec(45, 54), module, PolyrhythmClock::CLOCK_TOGGLE_PARAM));
         addParam(createParamCentered<PurpleKnob>(Vec(45, 76.7), module, PolyrhythmClock::BPM_PARAM));
-        // external clock
+        addParam(createParamCentered<TinyPurpleButton>(Vec(45, 54), module, PolyrhythmClock::CLOCK_TOGGLE_PARAM));
+        addChild(createLight<SmallLight<JeremyAquaLight>>(Vec(45 - 3.21, 54 - 3.21), module, PolyrhythmClock::TOGGLE_LIGHT));
+        // inputs
+        addInput(createInputCentered<TinyPJ301M>(Vec(19.9, 76.7), module, PolyrhythmClock::RESET_INPUT));
         addInput(createInputCentered<TinyPJ301M>(Vec(70.1, 76.7), module, PolyrhythmClock::EXT_CLOCK_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(19.9, 195.8), module, PolyrhythmClock::TUPLET1_RHYTHM_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(70.1, 195.8), module, PolyrhythmClock::TUPLET1_DUR_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(19.9, 271.1), module, PolyrhythmClock::TUPLET2_RHYTHM_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(70.1, 271.1), module, PolyrhythmClock::TUPLET2_DUR_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(19.9, 344.3), module, PolyrhythmClock::TUPLET3_RHYTHM_INPUT));
+        addInput(createInputCentered<TinyPJ301M>(Vec(70.1, 344.3), module, PolyrhythmClock::TUPLET3_DUR_INPUT));
 
         // tuplet 1
         addParam(createParamCentered<BlueInvertKnob>(Vec(19.9, 173.6), module, PolyrhythmClock::TUPLET1_RHYTHM_PARAM));
@@ -358,10 +425,10 @@ struct PolyrhythmClockWidget : ModuleWidget {
         addParam(createParamCentered<RedInvertKnob>(Vec(70.1, 322.1), module, PolyrhythmClock::TUPLET3_DUR_PARAM));
         addParam(createParamCentered<TinyRedKnob>(Vec(45, 322.7), module, PolyrhythmClock::TUPLETS_RAND_PARAM+2));
 
-        addOutput(createOutputCentered<PJ301MPort>(Vec(45, 119.8), module, PolyrhythmClock::MASTER_PULSE_OUTPUT));
-        addOutput(createOutputCentered<TinyPJ301M>(Vec(45, 195.8), module, PolyrhythmClock::TUPLET1_OUTPUT));
-        addOutput(createOutputCentered<TinyPJ301M>(Vec(45, 271.1), module, PolyrhythmClock::TUPLET2_OUTPUT));
-        addOutput(createOutputCentered<TinyPJ301M>(Vec(45, 344.3), module, PolyrhythmClock::TUPLET3_OUTPUT));
+        addOutput(createOutputCentered<PJ301MPurple>(Vec(45, 119.8), module, PolyrhythmClock::MASTER_PULSE_OUTPUT));
+        addOutput(createOutputCentered<TinyPortBlue>(Vec(45, 195.8), module, PolyrhythmClock::TUPLET1_OUTPUT));
+        addOutput(createOutputCentered<TinyPortAqua>(Vec(45, 271.1), module, PolyrhythmClock::TUPLET2_OUTPUT));
+        addOutput(createOutputCentered<TinyPortRed>(Vec(45, 344.3), module, PolyrhythmClock::TUPLET3_OUTPUT));
     }
 
     void appendContextMenu(Menu *menu) override {